changed perms, edited vimrc, and added .gitignore
for all the files in .bin, the remissions have been fixed to include execution.
the .gitignore file existed previous to this commit, but has just now been
edited to not include itself. as for the vimrc, removed the automatic fold
markers, toyed a bit with the remaining comments, added J K movement.